requirements.txt
selenium==4.9.1
instaloader==4.14.1
openai==0.27.8
pandas==2.0.3
gspread==5.7.1
oauth2client==4.1.3
requests==2.31.0
pyyaml==6.0
python-dotenv==1.0.0
